The list of languages is now built based on the existing gettext catalogs, rather than the Android resource files.
More specifically, this is the case for both the 'import' and 'export' commands. The 'init' command still works off the list of languages on the Android side. Internally, each command can use whatever makes more sense.

However, after some thought, for 'import' and 'export' it is a more correct solution to look for existing .po files, which are ultimately going to be the authoritative version of a translation.

In particular this changes makes things much easier for people who don't want to even keep the language-specific XML resources in version control.