Add support for handling formatted strings
Gettext supports the concept of allowing formatted strings and even
does some checking to make sure the translator didn't mess with the
formatting characters. I'm not sure Bable does anything with it, but it
may be useful in the future or with regular gettext tools to check
formatting before attempting to import.

If the formatting errors aren't caught before importing, it will crash
the Android application in the language that has a bad translation.