Ensure arrays are complete even if po is not.
If the gettext catalog is missing messages to complete an array,
ignore empty items in it's stead, when we can. If messages at
the end of an array are missing, we won't know about it.

This change makes the plural implementation easier, but is also
the correct behavior on it's own.