Support plurals. Fixes #10, #18.
This converts between Android's CLDR plurals, and the gettext
plural system. To do so, it needs to reproduce which quantities
Android uses for a language, and which it ignores (as the user
may define arbitrary quantities in Android even if they are not
supported, and cannot be reproduced in gettext).

Babel's builtin copy of the CLDR data is used to reproduce the
Android behavior. Should this turn out to be unsufficient due to
version differences, we would have to switch to the same CLDR
dataset that Android uses).

This implementation required touching various parts:

- po2xml() and xml2po() now need to know which language they are
  processing in order to resolve plurals. In particular, the
  interface of xml2po() has changed and it now no longer can be
  given a filename.

- The CLDR data for each language is loaded early, which causes
  the language code to be validated. Before, this only happend
  during export. The new behavior is more consistent.

So far, plural support has no test (and therefore a number of
bugs).