Respects explicit formatted="false" tag in strings
It is valid for Android XML files to have strings that disregard
the formatting characters (like percent) by setting the tag
formatted="false", such as:

   <string name="discount" formatted="false">50% off</string>

This tag was previously not being carried through an export/import
cycle. Now it is tracked properly. In the po file we will store
the flag "no-c-format" to mean formatted="false" in the XML.

Note: this is only implemented for <string> at this time, not
<plurals> and <string-array>.

Closes #24